Refactor: chat API 응답 DTO의 사용자 ID 타입을 Long으로 통일

+++ Back/petPlace/src/main/java/com/minjeok4go/petplace/feed/service/FeedService.java
@@ -88,7 +88,7 @@ public class FeedService {
                         : null)
                 .feedId(comment.getFeed().getId())
                 .content(comment.getContent())
-                .userId(comment.getUserId())
+                .userId(comment.getUserId().longValue())
                 .userNick(comment.getUserNick())
                 .userImg(comment.getUserImg())
                 .createdAt(comment.getCreatedAt())
